pyproject.toml
[project]
name = "impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dearcygui>=0.1.7",
    "pycryptodome>=3.23.0",
    "pyinstaller>=6.17.0",
    "pytest>=9.0.1",
    "rsa>=4.9.1",
]
